refactor(mirroring): move to heliotrope
