+++ heliotrope/server.py
@@ -11,7 +11,7 @@ from tortoise.contrib.sanic import register_tortoise
 import heliotrope
 from heliotrope.routes import heliotrope_routes
 from heliotrope.utils.requester import HitomiRequester
-from heliotrope.utils.tasks.mirroring import Mirroring
+from heliotrope.mirroring import Mirroring
 from heliotrope.utils.typed import Heliotrope
 
 heliotrope_app = Sanic("heliotrope")
